converted all floats to doubles

+++ src/cli.cpp
@@ -45,7 +45,7 @@ void validate_args(
             if      (arg_types[index] == "string") (std::string(args[i]));
             else if (arg_types[index] == "char")   char test = (char(args[i][0]));
             else if (arg_types[index] == "int")    std::stoi(args[i]);
-            else if (arg_types[index] == "float")  std::stof(args[i]);
+            else if (arg_types[index] == "double")  std::stof(args[i]);
             else if (arg_types[index] == "path")   open_file(std::string(args[i]));
         }
         catch(...)
